requirements.txt
aiogram==2.25.2
requests==2.31.0
Pillow==10.3.0